Added Required Repos and IServices, Services(ExamStatus,Exam)

+++ src/main/java/com/oea/online_exam_app/Models/ExamStatus.java
@@ -26,7 +26,7 @@ public class ExamStatus {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "exam_status_id")
-    private Long examStatusId;
+    private int examStatusId;
 
     @Column(name = "exam_status_txt", nullable = false)
     private String examStatusText;
